+++ demos/chinook/src/test/java/is/codion/framework/demos/chinook/domain/ChinookTest.java
@@ -28,7 +28,6 @@ import is.codion.framework.domain.entity.test.EntityTestUnit;
 
 import org.junit.jupiter.api.Test;
 
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 
@@ -124,7 +123,7 @@ public class ChinookTest extends EntityTestUnit {
 	protected Entity initializeTestEntity(EntityType entityType, Map<ForeignKey, Entity> foreignKeyEntities) {
 		Entity testEntity = super.initializeTestEntity(entityType, foreignKeyEntities);
 		if (entityType.equals(Album.TYPE)) {
-			testEntity.put(Album.TAGS, new HashSet<>(asList("tag_one", "tag_two")));
+			testEntity.put(Album.TAGS, asList("tag_one", "tag_two"));
 		}
 
 		return testEntity;
@@ -134,7 +133,7 @@ public class ChinookTest extends EntityTestUnit {
 	protected void modifyEntity(Entity testEntity, Map<ForeignKey, Entity> foreignKeyEntities) {
 		super.modifyEntity(testEntity, foreignKeyEntities);
 		if (testEntity.entityType().equals(Album.TYPE)) {
-			testEntity.put(Album.TAGS, new HashSet<>(asList("tag_one", "tag_two", "tag_three")));
+			testEntity.put(Album.TAGS, asList("tag_one", "tag_two", "tag_three"));
 		}
 	}
 }