+++ demos/manual/src/main/java/is/codion/framework/demos/manual/common/demo/ApplicationModel.java
@@ -19,8 +19,8 @@
 package is.codion.framework.demos.manual.common.demo;
 
 import is.codion.common.value.Value;
+import is.codion.common.value.ValueList;
 import is.codion.common.value.ValueObserver;
-import is.codion.common.value.ValueSet;
 import is.codion.swing.common.model.component.combobox.ItemComboBoxModel;
 
 import javax.swing.ComboBoxModel;
@@ -34,10 +34,11 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.List;
 
 import static is.codion.common.item.Item.item;
 import static is.codion.common.value.Value.value;
-import static is.codion.common.value.ValueSet.valueSet;
+import static is.codion.common.value.ValueList.valueList;
 import static is.codion.swing.common.model.component.combobox.ItemComboBoxModel.itemComboBoxModel;
 import static java.lang.Thread.setDefaultUncaughtExceptionHandler;
 import static java.util.Arrays.asList;
@@ -46,7 +47,7 @@ import static java.util.Arrays.asList;
 // tag::demoModelImport[]
 import static is.codion.common.item.Item.item;
 import static is.codion.common.value.Value.value;
-import static is.codion.common.value.ValueSet.valueSet;
+import static is.codion.common.value.ValueList.valueList;
 import static is.codion.swing.common.model.component.combobox.ItemComboBoxModel.itemComboBoxModel;
 import static java.lang.Thread.setDefaultUncaughtExceptionHandler;
 import static java.util.Arrays.asList;
@@ -72,7 +73,7 @@ public final class ApplicationModel {
 	private final Value<Integer> integerSpinValue = value();
 	private final Value<Integer> integerSelectionValue = value();
 	private final Value<String> itemSpinValue = value();
-	private final ValueSet<String> stringListValue = valueSet();
+	private final ValueList<String> stringListValue = valueList();
 	private final Value<String> messageValue = value();
 
 	private final Collection<Value<?>> values = asList(
@@ -168,7 +169,7 @@ public final class ApplicationModel {
 		return stringSelectionValue;
 	}
 
-	public ValueSet<String> stringListValueSet() {
+	public Value<List<String>> stringListValues() {
 		return stringListValue;
 	}