+++ changelog.md
@@ -4,10 +4,13 @@ Codion Change Log
 ## 0.17.42
 ### is.codion.common.core
 - Values added, extended by ValueSet, related refactoring.
+- ValueList added.
 ### is.codion.framework.domain
 - DefaultEntity refactored.
 - TypeReference now throws exception if used to specify a non-parameterized type.
 - DefaultDomain.add() now varargs based, related refactoring.
+### is.codion.swing.common.ui
+- ListBuilder migrated from using ValueSet to ValueList.
 
 ## 0.17.41
 ### is.codion.common.core